Week 4: Day 2
(The Trifecta) - Done
Added new safety measures in regard to the size of ice creams.
This includes an auto sizing option on creation that uses a new dictionary  the IceCreamSizes enum has now.
A new property method to retrieve the number of flavors in an icecream was made.
Tests were made for all the new features.

Notes:
I'm finished with everything for this week's assignment. Any more additions past today are either tweaks or additions likely to form a comprehensible API.