Week 4: Day 3 - 4
(The Trifecta) - Done

Notes:
WOOOOH! I have done it. My website is now able to make a call to my API and get all the item information. It is then able to procedurally create options for the user to choose. This works *perfectly*. Some more changes on the front-end on the web can make this look beautiful.